Rename file
Switch name to english ("Matrizen" is German ;) ).
