+++ module/data/item/_module.mjs
@@ -11,7 +11,7 @@ import KnackKarmaData from "./knack-karma.mjs";
 import KnackManeuverData from "./knack-maneuver.mjs";
 import ManeuverData from "./maneuver.mjs";
 import MaskData from "./mask.mjs";
-import MatrixData from "./matrizen.mjs"
+import MatrixData from "./matrices.mjs"
 import NamegiverData from "./namegiver.mjs";
 import PathData from "./path.mjs";
 import PoisonDiseaseData from "./poison-disease.mjs";
